+++ src/main/java/com/kcc/banking/domain/interest/mapper/InterestMapper.java
@@ -1,5 +1,8 @@
 package com.kcc.banking.domain.interest.mapper;
 
+import com.kcc.banking.domain.account_close.dto.request.PaymentStatus;
+import com.kcc.banking.domain.account_close.dto.request.RollbackPaymentStatus;
+import com.kcc.banking.domain.account_close.dto.response.InterestSum;
 import com.kcc.banking.domain.interest.dto.request.InterestCreate;
 import org.apache.ibatis.annotations.Mapper;
 
@@ -7,4 +10,10 @@ import org.apache.ibatis.annotations.Mapper;
 public interface InterestMapper {
 
     void createInterest(InterestCreate interestCreate);
+
+    InterestSum findInterestSum(String accountId);
+
+    int updatePaymentStatus(PaymentStatus paymentStatus);
+
+    int rollbackPaymentStatus(RollbackPaymentStatus rollbackPaymentStatus);
 }
