+++ src/main/resources/mapper/accountCloseMapper.xml
@@ -4,10 +4,6 @@
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 <mapper namespace="com.kcc.banking.domain.account_close.mapper.AccountCloseMapper">
 
-    <resultMap id="interestSumMap" type="InterestSum">
-        <id column="account_id" property="accountId" />
-        <result column="amount_sum" property="amountSum" />
-    </resultMap>
 
     <resultMap id="closeAccountMap" type="CloseAccount">
         <id column="account_id" property="accountId" />
@@ -22,16 +18,6 @@
         <result column="product_tax_rate" property="productTaxRate" />
     </resultMap>
 
-<!--    계좌번호로 미지급 이자,미지급 이자율 총합-->
-    <select id="findInterestSum" parameterType="java.lang.String" resultMap="interestSumMap">
-        SELECT a.id AS account_id,
-               sum(i.amount) AS amount_sum
-        FROM Account a
-            JOIN Interest i ON  i.acc_id = a.id
-        WHERE a.id = #{accountId}
-            AND i.payment_status = 'N'
-        GROUP BY a.id
-    </select>
 
 <!--    해지 대상 계좌 상태 조회-->
     <select id="findCloseAccount" parameterType="java.lang.String" resultMap="closeAccountMap">
@@ -59,37 +45,6 @@
         INSERT INTO TRADE (id, acc_id, registrant_id, branch_id, trade_date, amount, balance, trade_type, status, cash_indicator, description, registration_date, trade_number, version)
         VALUES (#{id}, #{accId}, #{registrantId}, #{branchId}, #{businessDay}, #{amount}, #{balance}, #{tradeType}, 'NOR', 'TRUE', #{description}, SYSDATE, trade_num_seq.nextval, 1)
     </insert>
-
-<!--    이자 테이블 상태 변경-->
-    <update id="updatePaymentStatus" parameterType="PaymentStatus">
-        UPDATE INTEREST
-            SET
-                branch_id = #{branchId},
-                payment_date = #{payDate},
-                payment_status = 'Y',
-                modifier_date = SYSDATE,
-                modifier_id = #{modifierId},
-                version = version + 1
-            WHERE
-                acc_id = #{accId}
-                AND
-                payment_status = 'N'
-    </update>
-
-    <update id="rollbackPaymentStatus" parameterType="RollbackPaymentStatus">
-        UPDATE INTEREST
-        SET
-            branch_id = #{branchId},
-            payment_date = NULL,
-            payment_status = 'N',
-            modifier_date = SYSDATE,
-            modifier_id = #{modifierId},
-            version = version + 1
-        WHERE
-            acc_id = #{accId}
-          AND
-            modifier_date = #{expireDate}
-    </update>
 
 <!--    계좌 상태 변경-->
     <update id="updateStatus" parameterType="AccountStatus">
