+++ src/main/webapp/resources/js/page/account-close.js
@@ -2,6 +2,7 @@ let accountData = {};
 
 $(document).ready(function () {
     $('#search-modal-account').on('hidden.bs.modal', function () {
+        console.log("모달 닫힘");
         getAccountDetail();
     });
